feat: immigrate home function to profile page.
