+++ packages/nextjs/app/feed/contentPage.tsx
@@ -40,17 +40,19 @@ const ContentPage: NextPage = function () {
   return (
     <>
       <PostDisplayContext.Provider value={{ compactDisplay: true, showPostMgmt: false, onProfile: false }}>
-        {data.map((post: TPost) => (
-          <div
-            key={post.id}
-            className="flex flex-col w-full p-4 gap-4 justify-start
-                      transition-all duration-300 ease-in-out
-                      hover:bg-gray-200 hover:shadow-lg hover:scale-105 cursor-pointer"
-            onClick={() => onClickViewPost(post.id)}
-          >
-            <Post key={post.id} postId={post.id} />
-          </div>
-        ))}
+        <div className="flex flex-col items-center w-full">
+          {data.map((post: TPost) => (
+            <div
+              key={post.id}
+              className="flex flex-col w-[95%] p-4 gap-4 justify-start
+                        transition-all duration-300 ease-in-out
+                        hover:bg-gray-200 hover:shadow-lg hover:scale-105 cursor-pointer"
+              onClick={() => onClickViewPost(post.id)}
+            >
+              <Post key={post.id} postId={post.id} />
+            </div>
+          ))}
+        </div>
       </PostDisplayContext.Provider>
     </>
   );
