+++ packages/nextjs/app/home/profile.tsx
@@ -1,25 +1,74 @@
-import { useEffect } from "react";
+import { useContext, useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
+import type { NextPage } from "next";
 import { Hex } from "viem";
 import { useAccount } from "wagmi";
 import Post from "~~/components/post/Post";
 import { PostDisplayContext } from "~~/contexts/Post";
+import { UserProfileContext } from "~~/contexts/UserProfile";
 import { useScaffoldReadContract } from "~~/hooks/scaffold-eth";
+import { useScaffoldWriteContract } from "~~/hooks/scaffold-eth/useScaffoldWriteContract";
 import { TPost, TUserProfile } from "~~/types/spotlight";
+import { notification } from "~~/utils/scaffold-eth";
 
 function LeftColumn() {
+  const router = useRouter();
   const { address: connectedAddress } = useAccount();
-  const { data: userProfile } = useScaffoldReadContract({
-    contractName: "Spotlight",
-    functionName: "getProfile",
-    args: [connectedAddress],
-    watch: true,
-  }) as {
-    data: TUserProfile | undefined;
-  };
   const shortenedUserAddress = connectedAddress
     ? connectedAddress.slice(0, 6) + "..." + connectedAddress.slice(-4)
     : "";
+
+  const [deleted, setDeleted] = useState(false);
+  const [isChangingUsername, setIsChangingUsername] = useState(false);
+  const [newUsername, setNewUsername] = useState<string>("");
+  const { writeContractAsync: writeSpotlightContractAsync } = useScaffoldWriteContract("Spotlight");
+  const { userProfile } = useContext(UserProfileContext);
+
+  useEffect(() => {
+    if (userProfile && userProfile.username === "") {
+      // They need to go register first...
+      router.push("/");
+    }
+  }, [userProfile, router]);
+
+  const handleDeleteProfile = async () => {
+    const confirmDelete = window.confirm("Are you sure you want to delete your profile?");
+    if (!confirmDelete) return;
+
+    try {
+      await writeSpotlightContractAsync({
+        functionName: "deleteProfile",
+      });
+      setDeleted(true);
+    } catch (e: any) {
+      console.error(e);
+      notification.error("Failed to delete profile");
+    }
+  };
+
+  const handleChangeUsername = async () => {
+    if (!newUsername) {
+      notification.error("Please enter a new username.");
+      return;
+    }
+
+    try {
+      await writeSpotlightContractAsync({
+        functionName: "updateUsername",
+        args: [newUsername],
+      });
+      notification.success("Username updated successfully!");
+      setIsChangingUsername(false); // Hide the form once successful
+    } catch (e: any) {
+      console.error(e);
+      notification.error("Failed to update username");
+    }
+  };
+
+  if (userProfile === undefined) {
+    // We cannot render anything
+    return;
+  }
 
   return (
     <div className="w-[100%] flex flex-col items-center">
@@ -29,8 +78,48 @@ function LeftColumn() {
         className="w-80 h-80 rounded-full object-cover ml-2"
       />
       <div className="flex w-[100%] flex-col pl-4">
-        <h1 className="text-4xl  text-left">{userProfile?.username}</h1>
+        <h1 className="text-left text-2xl font-bold text-gray-800">{userProfile?.username}</h1>
         <div>Address: {shortenedUserAddress}</div>
+        <div>Reputation: {Number(userProfile.reputation)} RPT</div>
+      </div>
+      <div className="flex flex-col justify-start items-center flex-grow-0 flex-shrink-0 sm:mt-4 mb-4">
+        {deleted ? (
+          <>Deleting profile...</>
+        ) : (
+          <>
+            {isChangingUsername ? (
+              // Render the form for changing username
+              <div className="w-full flex flex-col items-center">
+                <div className="mb-4 center">
+                  <input
+                    type="text"
+                    placeholder="Enter new username"
+                    className="input input-bordered"
+                    value={newUsername}
+                    onChange={e => setNewUsername(e.target.value)}
+                  />
+                </div>
+                <div className="flex justify-center space-x-4">
+                  <button className="btn btn-primary" onClick={handleChangeUsername}>
+                    Submit New Username
+                  </button>
+                  <button className="btn btn-error" onClick={() => setIsChangingUsername(false)}>
+                    Cancel
+                  </button>
+                </div>
+              </div>
+            ) : (
+              <div className="flex space-x-4 mt-4">
+                <button className="btn btn-secondary" onClick={() => setIsChangingUsername(true)}>
+                  Change Username
+                </button>
+                <button className="btn btn-error" onClick={handleDeleteProfile}>
+                  Delete Profile
+                </button>
+              </div>
+            )}
+          </>
+        )}
       </div>
     </div>
   );
